move all UI related files into own ui package to separate from run configurations.
