+++ skaffold/src/main/kotlin/com/google/container/tools/skaffold/run/ui/SkaffoldSingleRunSettingsEditor.kt
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package com.google.container.tools.skaffold.run
+package com.google.container.tools.skaffold.run.ui
 
 import com.google.common.annotations.VisibleForTesting
 import com.google.container.tools.skaffold.message
+import com.google.container.tools.skaffold.run.SkaffoldSingleRunConfiguration
 import javax.swing.JCheckBox
 
 /**
