+++ skaffold/src/test/kotlin/com/google/container/tools/skaffold/run/BaseSkaffoldSettingsEditorTest.kt
@@ -18,6 +18,7 @@ package com.google.container.tools.skaffold.run
 
 import com.google.common.truth.Truth.assertThat
 import com.google.container.tools.skaffold.SkaffoldFileService
+import com.google.container.tools.skaffold.run.ui.BaseSkaffoldSettingsEditor
 import com.google.container.tools.test.ContainerToolsRule
 import com.google.container.tools.test.TestService
 import com.google.container.tools.test.UiTest
@@ -54,7 +55,8 @@ class BaseSkaffoldSettingsEditorTest {
     @Before
     fun setUp() {
         EdtTestUtil.runInEdtAndWait(ThrowableRunnable {
-            baseSkaffoldSettingsEditor = BaseSkaffoldSettingsEditor("test")
+            baseSkaffoldSettingsEditor =
+                BaseSkaffoldSettingsEditor("test")
             // calls getComponent() to initialize UI first, IDE flow.
             baseSkaffoldSettingsEditor.component
         })
