+++ skaffold/src/test/kotlin/com/google/container/tools/skaffold/run/SkaffoldFilesComboBoxTest.kt
@@ -18,6 +18,7 @@ package com.google.container.tools.skaffold.run
 
 import com.google.common.truth.Truth.assertThat
 import com.google.container.tools.skaffold.SkaffoldFileService
+import com.google.container.tools.skaffold.run.ui.SkaffoldFilesComboBox
 import com.google.container.tools.test.ContainerToolsRule
 import com.google.container.tools.test.TestService
 import com.google.container.tools.test.UiTest
